factor out side panel toggle
