+++ lightly_studio_view/src/lib/components/SidePanelTabs/useSidePanelTabs.ts
@@ -1,0 +1,22 @@
+import { useGlobalStorage, usePostHog } from '$lib/hooks';
+
+type PanelType = Parameters<ReturnType<typeof useGlobalStorage>['setActivePanel']>[0];
+
+interface UseSidePanelTabsParams {
+    collectionId: string;
+}
+
+export function useSidePanelTabs({ collectionId }: UseSidePanelTabsParams) {
+    const { activePanel, setActivePanel } = useGlobalStorage();
+    const { trackEvent } = usePostHog();
+
+    function toggle(activePanel: PanelType | 'none', panel: PanelType) {
+        const nextPanel = activePanel === panel ? 'none' : panel;
+        setActivePanel(nextPanel);
+        if (nextPanel !== 'none') {
+            trackEvent('tool_panel_opened', { collection_id: collectionId, panel_type: nextPanel });
+        }
+    }
+
+    return { activePanel, toggle };
+}
